refactor(onboard): move onboarding routes
