+++ src/treg/routers/onboard.py
@@ -1,0 +1,99 @@
+"""HTTP routes for first-run team onboarding."""
+
+from fastapi import APIRouter, Depends, HTTPException
+from pydantic import BaseModel
+from sqlalchemy.ext.asyncio import AsyncSession
+from sqlmodel import select
+
+from .. import demo as demo_seed
+from ..db import get_session
+from ..domain.identity.access import (
+    Caller,
+    _norm_email,
+    _require_can_register,
+    require_identity,
+    require_member,
+)
+from ..models import Invite, Org, User
+from .orgs import _require_admin_of
+
+
+# app is the APIRouter alias so mechanically moved @app decorators stay byte-identical.
+app = APIRouter()
+
+
+class OnboardIn(BaseModel):
+    team_name: str = "Acme Design"
+
+
+@app.post("/onboard/demo")
+async def onboard_demo(
+    body: OnboardIn | None = None,
+    user: User = Depends(require_identity),
+    db: AsyncSession = Depends(get_session),
+) -> dict:
+    """Seed a sandbox team owned by the caller — fake teammates (one per role) + a working `echo`
+    tool + sample activity — so a brand-new user can feel the product immediately. Idempotent
+    (reuses an existing demo team); marks the caller onboarded. Same seed for dashboard + CLI."""
+    return await demo_seed.provision(db, user, (body.team_name if body else "Acme Design"))
+
+
+@app.post("/onboard/skip")
+async def onboard_skip(
+    user: User = Depends(require_identity), db: AsyncSession = Depends(get_session)
+) -> dict:
+    """Dismiss onboarding without seeding — so it's never auto-offered again."""
+    user.onboarded = True
+    await db.commit()
+    return {"onboarded": True}
+
+
+@app.post("/onboard/reset")
+async def onboard_reset(
+    user: User = Depends(require_identity), db: AsyncSession = Depends(get_session)
+) -> dict:
+    """Remove the caller's demo team(s) + demo teammates from their real teams — a clean exit."""
+    return await demo_seed.reset(db, user)
+
+
+onboard_entry_router = app
+
+# The second router preserves the later attachment point after the landing-sandbox routes.
+app = APIRouter()
+
+
+class TeammateIn(BaseModel):
+    email: str
+
+
+@app.post("/onboard/seed-tool")
+async def onboard_seed_tool(
+    caller: Caller = Depends(require_member), db: AsyncSession = Depends(get_session)
+) -> dict:
+    """Pre-seed the working `echo` tool into the caller's active team so the no-key call in the
+    dashboard onboarding just works (the user builds the team + invites by hand; the tool is on us)."""
+    _require_can_register(caller)
+    org = await db.get(Org, caller.org_id)
+    if org is None:
+        raise HTTPException(status_code=404, detail="org not found")
+    return await demo_seed.seed_tool(db, org, caller.email)
+
+
+@app.post("/onboard/accept-teammate")
+async def onboard_accept_teammate(
+    body: TeammateIn, caller: Caller = Depends(require_member), db: AsyncSession = Depends(get_session)
+) -> dict:
+    """Auto-accept the fake teammate the user just invited during onboarding, so it lands in the
+    roster instantly (they feel the invite, then see the loop close). Admin+ only, demo email only."""
+    _require_admin_of(caller.org_id, caller)
+    email = _norm_email(body.email)
+    if not email.endswith("@" + demo_seed.DEMO_DOMAIN):
+        raise HTTPException(status_code=400, detail="onboarding auto-accept is for demo teammates only")
+    inv = (await db.execute(select(Invite).where(
+        Invite.org_id == caller.org_id, Invite.email == email, Invite.status == "pending"))).scalar_one_or_none()
+    if inv is None:
+        raise HTTPException(status_code=404, detail="no pending invite for that email")
+    return await demo_seed.accept_demo_invite(db, caller.org_id, inv)
+
+
+onboard_teammate_router = app
